requirements.txt
EasyDict
Pillow
pyyaml
sklearn
numpy==1.16.1